Add update cat

+++ app/src/main/java/com/android/zore3x/artoscats/activity/CatsListActivity.kt
@@ -36,7 +36,7 @@ class CatsListActivity : MvpAppCompatActivity(), ICatListView {
             startActivity(EditableCatActivity.getIntent(applicationContext))
         }
         adapter.setOnClickListener(object : CatAdapter.CatClickListener {
-            override fun OnClick(view: View, position: Int) {
+            override fun OnClick(view: View, position: Long) {
                 startActivity(EditableCatActivity.getIntent(applicationContext, position))
             }
         })
